Move `h.views.api_auth` -> `h.views.api.auth`
